Added server.js with export and import
