+++ index.js
@@ -1,28 +1,4 @@
-// implement your API here
-
-const express = require('express');
-
-const db = require('./data/db.js');
-
-const server = express();
-
-// CRUD
-
-// Create 
-server.post('', (req, res) => {
-});
-
-// Read 
-server.get('', (req, res) => {
-});
-
-// Update 
-server.put('', (req, res) => {
-});
-
-// Delete 
-server.delete('', (req, res) => {
-});
+const server = require('./server.js');
 
 server.listen(4000, () => {
     console.log('\n** API up and running on port 4k **');
