+++ server.js
@@ -1,0 +1,27 @@
+// implement your API here
+
+const express = require('express');
+
+const db = require('./data/db.js');
+
+const server = express();
+
+// CRUD
+
+// Create 
+server.post('', (req, res) => {
+});
+
+// Read 
+server.get('', (req, res) => {
+});
+
+// Update 
+server.put('', (req, res) => {
+});
+
+// Delete 
+server.delete('', (req, res) => {
+});
+
+module.exports = server;
